flashrom: Fix compilation of r3797 with gcc-4.3.2
Thanks to Niels Ole Salscheider for the problem report.

Acked-by: Peter Stuge <peter@stuge.se>


git-svn-id: svn://svn.coreboot.org/coreboot/trunk@3798 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ util/flashrom/flashrom.c
@@ -250,7 +250,7 @@ int main(int argc, char *argv[])
 {
 	uint8_t *buf;
 	unsigned long size;
-	size_t erasedbytes;
+	uint32_t erasedbytes;
 	FILE *image;
 	/* Probe for up to three flash chips. */
 	struct flashchip *flash, *flashes[3];
